basic gameplay and ammo interactions
basic gameplay mechanics have been introduced and interactions between boss and player ammo types have been developed. Boss can lose health and eventually die, which would lead to the boss getting harder and the next wave starting. That being said, the bombs the boss drops do not seem to register when collided with to deal damage to the player