Finished base game
The bombs now collide properly with the player, they also remain on the ground for a duration of time before being destroyed as an additional obstacle. When the boss is defeated the boss despawns for a short duration before being reactivated with its stats boosted and the next round begins.